Reads and merges all comexstat files with information about NCM

Reads and merges all comexstat files with information about NCM

Usage

ncms()

Value

data.frame/tibble

Details

reads the following files: ncm, ncm_cgce, ncm_cuci, ncm_isic, ncm_unidade then joins them into a single tibble

Examples

ncms()
#> # A tibble: 13,677 × 51
#>    co_ncm   co_unid co_sh6 co_ppe co_ppi co_fat_agreg co_cuci_item co_cgce_n3
#>    <chr>    <chr>   <chr>  <chr>  <chr>  <chr>        <chr>        <chr>     
#>  1 49011000 10      490110 3521   3521   03           89215        322       
#>  2 49019100 10      490191 3521   3521   03           89216        322       
#>  3 49019900 10      490199 3521   3521   03           89219        322       
#>  4 49021000 10      490210 3521   3521   03           89221        322       
#>  5 52102910 10      521029 3840   3840   03           65251        240       
#>  6 52102990 10      521029 3840   3840   03           65251        240       
#>  7 52103100 10      521031 3840   3840   03           65252        240       
#>  8 52103200 10      521032 3840   3840   03           65252        240       
#>  9 52103900 10      521039 3840   3840   03           65252        240       
#> 10 52104100 10      521041 3840   3840   03           65253        240       
#> # ℹ 13,667 more rows
#> # ℹ 43 more variables: co_siit <chr>, co_isic_classe <chr>,
#> #   co_exp_subset <chr>, no_ncm_por <chr>, no_ncm_esp <chr>, no_ncm_ing <chr>,
#> #   no_cgce_n3 <chr>, no_cgce_n3_ing <chr>, no_cgce_n3_esp <chr>,
#> #   co_cgce_n2 <chr>, no_cgce_n2 <chr>, no_cgce_n2_ing <chr>,
#> #   no_cgce_n2_esp <chr>, co_cgce_n1 <chr>, no_cgce_n1 <chr>,
#> #   no_cgce_n1_ing <chr>, no_cgce_n1_esp <chr>, no_cuci_item <chr>, …